[Configuration] Reorganize clues into separate files (#198)
* Split rules into separate files

* Split tests

* Move tests

* Fix tests

* Organize

* Rename

* Organize

* Clean up paths

* Clean up

* Clean up
