+++ package.json
@@ -55,8 +55,9 @@
   ],
   "jest": {
     "moduleNameMapper": {
-      "^./clues.js$": "<rootDir>/src/clues.ts",
-      "^./rules.js$": "<rootDir>/src/rules.ts"
+      "./clues.js$": "<rootDir>/src/clues.ts",
+      "^./clues/(.*)/index.js$": "<rootDir>/src/clues/$1/index.ts",
+      "../index.js$": "<rootDir>/src/index.ts"
     }
   }
 }
