+++ src/clues/avoid-emoji/__tests__/index.test.ts
@@ -1,0 +1,9 @@
+import { checkEmoji } from "..";
+
+test("checkEmoji", () => {
+  expect(checkEmoji("An orange 🐈.")).toMatchInlineSnapshot(`
+      Array [
+        "Replace 🐈 in alt text with descriptive text (https://tinyurl.com/yylrxrus).",
+      ]
+    `);
+});
