+++ src/clues/contains/__tests__/index.test.ts
@@ -1,0 +1,9 @@
+import { checkClue } from "../../../clues.js";
+
+test("contains", () => {
+  expect(checkClue("A screenshot of a dog.")).toMatchInlineSnapshot(`
+      Array [
+        "Alt text should not contain \\"screenshot of\\" (https://tinyurl.com/y3v3jgux).",
+      ]
+    `);
+});
