+++ src/clues/end-punctuation/__tests__/index.test.ts
@@ -1,0 +1,21 @@
+import { checkPunctuation } from "..";
+
+test("checkPunctuation", () => {
+  expect(checkPunctuation("a large black dog")).toMatchInlineSnapshot(`
+      Array [
+        "Alt text should end with punctuation (https://tinyurl.com/y5krn3ny).",
+      ]
+    `);
+  expect(checkPunctuation("a large black dog.")).toMatchInlineSnapshot(
+    [],
+    `Array []`
+  );
+  expect(checkPunctuation("a large black dog!")).toMatchInlineSnapshot(
+    [],
+    `Array []`
+  );
+  expect(checkPunctuation("a large black dog?")).toMatchInlineSnapshot(
+    [],
+    `Array []`
+  );
+});
