+++ src/clues/image-link/index.ts
@@ -1,0 +1,16 @@
+import { Rule } from "../../clues.js";
+
+export default function imageLink(): Rule {
+  return {
+    heading: "Image is link",
+    docs: "https://tinyurl.com/yxhq2k5w",
+    suggestion: () =>
+      `Images inside a link tag require alt text that describes the purpose of the link`,
+    rationale:
+      "Images inside a link tag require alt text that describes the purpose of the link.",
+    source: ["https://axesslab.com/alt-texts/#images-in-links"],
+    ok: '`<a href="https://github.com/double-great"><img src="logo.png" alt="double great on GitHub"></a>`',
+    notOk:
+      '`<a href="https://github.com/double-great"><img src="logo.png" alt="double great logo"></a>`',
+  };
+}
