+++ src/clues/not-only-space/__tests__/index.test.ts
@@ -1,0 +1,9 @@
+import { checkOnlySpace } from "..";
+
+test("checkOnlySpace", () => {
+  expect(checkOnlySpace(" ")).toMatchInlineSnapshot(`
+      Array [
+        "Alt text should not be a single space (https://tinyurl.com/y2o7uctu).",
+      ]
+    `);
+});
