+++ src/clues/start-with/__tests__/index.test.ts
@@ -1,0 +1,9 @@
+import { checkClue } from "../../../clues.js";
+
+test("startWith", () => {
+  expect(checkClue("spacer image.")).toMatchInlineSnapshot(`
+      Array [
+        "Alt text should not start with \\"spacer\\" (https://tinyurl.com/y5y98ygu).",
+      ]
+    `);
+});
